+++ examples/cross-platform/glue/prg.rs
@@ -13,3 +13,12 @@ fn main() {
 
 #[cfg(not(any(target_arch = "wasm32", target_os = "android")))]
 mod _glue_entry_point {}
+
+#[cfg(target_arch = "wasm32")]
+fn main() {
+    const LOG: devout::Tag = devout::Tag::new("LOG");
+
+    devout::log!(LOG, "oof");
+    std::thread::park();
+    devout::log!(LOG, "da oof");
+}