+++ examples/slices.rs
@@ -1,5 +1,5 @@
 use core::task::Poll;
-use pasts::{Loop, Task};
+use pasts::{Executor, Loop, Task};
 
 type Exit = ();
 
@@ -22,8 +22,10 @@ async fn run() {
     Loop::new(&mut state)
         .poll(|s| &mut s.tasks, State::completion)
         .await;
+
+    std::process::exit(0)
 }
 
 fn main() {
-    pasts::block_on(run())
+    Executor::new().cycle(run())
 }